perplexity/requirements.txt
flask
flask-cors
google-genai
PyPDF2
